ver 0.838
新增:
 更多關於裝備重量的東西

+++ Dungeon/unit.h
@@ -19,8 +19,8 @@ struct unit {
 	
 	
 	//ª±®a
-	int maxWeight;
-	int weight;
+	float maxWeight;
+	float weight;
 	char name[100];
 	bool inventoryMode = false;
 	//¼Ä¤H
